refactor: move gradient to Header Folder
