+++ src/components/Header/Gradient/index.tsx
