+++ src/components/Header/Gradient/styles.ts
