+++ src/components/Header/index.tsx
@@ -10,7 +10,7 @@ import { BancoInter } from '../../assets/images';
 import Button from '../Button';
 import AccountDropdown from './AccountDropdown';
 import { User } from './Header';
-import Gradient from '../Gradient';
+import Gradient from './Gradient';
 
 const Header: React.FC = () => {
   const user: User = { name: 'Isaque Igor' };
